+++ app/src/main/java/com/example/shoppingcart/models/Product.java
@@ -18,7 +18,7 @@ public class Product {
     private String name;
 
     @SerializedName("price")
-    private double price;
+    private int price;
 
     private boolean isAvailable = true;
 
@@ -34,7 +34,7 @@ public class Product {
     @SerializedName("genre")
     private String genre;
 
-    public Product(String id, String name, double price, String imageUrl, String description, String storeId, String genre) {
+    public Product(String id, String name, int price, String imageUrl, String description, String storeId, String genre) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -64,7 +64,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
